Deleted Components/APIs folder and Astronomy.jsx, moved TriggerMyMission.jsx to Components folder root
